use u32 type rather u64 to record total amount of posts

+++ src/persister.rs
@@ -245,12 +245,12 @@ impl Persister {
         }
     }
 
-    pub async fn query_db_total_num(&self) -> Result<u64> {
+    pub async fn query_db_total_num(&self) -> Result<u32> {
         Ok(
-            sqlx::query_as::<Sqlite, (i64,)>("SELECT COUNT(1) FROM posts WHERE favorited")
+            sqlx::query_as::<Sqlite, (u32,)>("SELECT COUNT(1) FROM posts WHERE favorited")
                 .fetch_one(self.db_pool.as_ref().unwrap())
                 .await?
-                .0 as u64,
+                .0 as u32,
         )
     }
 }
